Stop caching, for now
Caching tools is recommended if you plan to "switch back and forth
between versions of a tool, or save a tool between runs for self-hosted
runners." Neither of these situations seem necessary for dokube at the
moment, not to mention caching wasn't really even implemented correctly
in the first place, as the action never hunts for existing versions of
the tool in the cache, it just keeps downloading and adding to cache.

Hopefully this will solve a problem raised in #14.

https://github.com/actions/toolkit/tree/master/packages/tool-cache#cache

.github/workflows/test.yml |   1 +
dist/index.js              | 266 ++++++++++++++++++++++++++++++-------
index.js                   |   7 +-
3 files changed, 221 insertions(+), 53 deletions(-)